+++ lib/config/router.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/route_manager.dart';
 import 'package:mi_utem/config/routes.dart';
+import 'package:mi_utem/controllers/asignatura_controller.dart';
+import 'package:mi_utem/screens/asignatura_screen.dart';
 import 'package:mi_utem/screens/asignaturas_screen.dart';
 import 'package:mi_utem/screens/calculadora_notas_screen.dart';
 import 'package:mi_utem/screens/credencial_screen.dart';
@@ -65,7 +67,12 @@ final pages = <GetPage>[
     name: Routes.asignaturas,
     page: () => AsignaturasScreen(),
     middlewares: [OnlyAuthMiddleware()],
+    binding: AsignaturaBinding(),
   ),
+  GetPage(
+    name: Routes.asignatura,
+    page: () => AsignaturaScreen(),
+  )
 ];
 
 class OnlyAuthMiddleware extends GetMiddleware {