+++ lib/controllers/notification_controller.dart
@@ -6,8 +6,8 @@ import 'package:awesome_notifications_fcm/awesome_notifications_fcm.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mi_utem/config/routes.dart';
+import 'package:mi_utem/controllers/asignatura_controller.dart';
 import 'package:mi_utem/models/asignatura.dart';
-import 'package:mi_utem/screens/asignatura_screen.dart';
 import 'package:mi_utem/services/analytics_service.dart';
 
 class NotificationController {
@@ -56,10 +56,8 @@ class NotificationController {
         );
         final asignatura =
             Asignatura.fromJson(jsonDecode(asignaturaJsonString));
-        Get.to(
-          () => AsignaturaScreen(asignatura: asignatura),
-          routeName: Routes.asignatura,
-        );
+        AsignaturaController.to.asignatura.value = asignatura;
+        Get.toNamed(Routes.asignatura);
       }
     }
   }